backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.11
PyJWT>=2.8.0
python-dotenv>=1.0.0
pydantic>=2.12.0
